+++ sdk/assemblyscript/src/assembly/neo4j.ts
@@ -74,7 +74,24 @@ export class Record {
       isString<T>() ||
       idof<T>() === idof<Node>() ||
       idof<T>() === idof<Relationship>() ||
-      idof<T>() === idof<Path>()
+      idof<T>() === idof<Path>() ||
+      idof<T>() === idof<DynamicMap>() ||
+      idof<T>() === idof<string[]>() ||
+      idof<T>() === idof<i8[]>() ||
+      idof<T>() === idof<i16[]>() ||
+      idof<T>() === idof<i32[]>() ||
+      idof<T>() === idof<i64[]>() ||
+      idof<T>() === idof<u8[]>() ||
+      idof<T>() === idof<u16[]>() ||
+      idof<T>() === idof<u32[]>() ||
+      idof<T>() === idof<u64[]>() ||
+      idof<T>() === idof<f32[]>() ||
+      idof<T>() === idof<f64[]>() ||
+      idof<T>() === idof<Date>() ||
+      idof<T>() === idof<JSON.Raw>() ||
+      idof<T>() === idof<JSON.Raw[]>() ||
+      idof<T>() === idof<Point2D>() ||
+      idof<T>() === idof<Point3D>()
     ) {
       for (let i = 0; i < this.Keys.length; i++) {
         if (this.Keys[i] == key) {
@@ -107,7 +124,16 @@ abstract class Entity {
   Props!: DynamicMap;
 
   getProperty<T>(key: string): T {
-    if (isInteger<T>() || isFloat<T>() || isBoolean<T>() || isString<T>()) {
+    if (
+      isInteger<T>() ||
+      isFloat<T>() ||
+      isBoolean<T>() ||
+      isString<T>() ||
+      idof<T>() === idof<Date>() ||
+      idof<T>() === idof<JSON.Raw>() ||
+      idof<T>() === idof<Point2D>() ||
+      idof<T>() === idof<Point3D>()
+    ) {
       return this.Props.get<T>(key);
     }
     throw new Error("Unsupported type.");
@@ -148,4 +174,48 @@ export class Path {
 
   @alias("Relationships")
   Relationships!: Relationship[];
+}
+
+
+@json
+export class Point2D {
+
+  @alias("X")
+  X!: f64;
+
+
+  @alias("Y")
+  Y!: f64;
+
+
+  @alias("SpatialRefId")
+  SpatialRefId!: u32;
+
+  String(): string {
+    return `Point{SpatialRefId=${this.SpatialRefId}, X=${this.X}, Y=${this.Y}}`;
+  }
+}
+
+
+@json
+export class Point3D {
+
+  @alias("X")
+  X!: f64;
+
+
+  @alias("Y")
+  Y!: f64;
+
+
+  @alias("Z")
+  Z!: f64;
+
+
+  @alias("SpatialRefId")
+  SpatialRefId!: u32;
+
+  String(): string {
+    return `Point{SpatialRefId=${this.SpatialRefId}, X=${this.X}, Y=${this.Y}, Z=${this.Z}}`;
+  }
 }