+++ CHANGELOG.md
@@ -13,6 +13,8 @@
 - fix: vector package should return generic type in computations [#628](https://github.com/hypermodeinc/modus/pull/628)
 - chore: Remove unused go package reference [#632](https://github.com/hypermodeinc/modus/pull/632)
 - feat: add neo4j support in modus [#636](https://github.com/hypermodeinc/modus/pull/636)
+- fix: support point types [#650](https://github.com/hypermodeinc/modus/pull/650)
+
 
 ## UNRELEASED - AssemblyScript SDK
 
@@ -21,6 +23,7 @@
 - feat: add neo4j support in modus [#636](https://github.com/hypermodeinc/modus/pull/636)
 - feat: Add `DynamicMap` type [#638](https://github.com/hypermodeinc/modus/pull/638)
 - fix: Dgraph AssemblyScript example query issues [#643](https://github.com/hypermodeinc/modus/pull/643)
+- fix: neo4j as type unmarshalling & support point types [#650](https://github.com/hypermodeinc/modus/pull/650)
 
 ## 2024-12-11 - CLI 0.15.0